+++ protocols/beammp.js
@@ -7,7 +7,10 @@ export default class beammp extends Core {
     master.options = this.options
     const masterState = await master.runOnceSafe()
     const servers = masterState.raw.servers
-    const server = servers.find(s => s.ip === this.options.address)
+    const sameIpServers = servers.filter(s => s.ip === this.options.address)
+    const sortedServers = sameIpServers.sort(s => parseInt(s.port))
+    // Cause some don't have ports specified, so we prioritize those who do
+    const server = sortedServers.find(s => s.ip === this.options.address && this.matchPort(s.port))
 
     if (!server) {
       throw new Error('Server not found in the master list')
@@ -29,5 +32,15 @@ export default class beammp extends Core {
 
     state.raw = server
     if ('version' in state.raw) state.version = state.raw.version
+  }
+
+  matchPort (givenPort) {
+    const parsedPort = parseInt(givenPort)
+    const port = this.options.port
+    if (!port || isNaN(parsedPort)) {
+      return true
+    }
+
+    return port === parsedPort
   }
 }